Add: exported configure() function, Update: LitNodeClient constructor to use configure()
