+++ src/utils/litNodeClient.js
@@ -26,6 +26,7 @@ import {
   canonicalResourceIdFormatter,
   canonicalUnifiedAccessControlConditionFormatter,
 } from "./crypto";
+import { configure } from "./lit";
 
 /**
  * @typedef {Object} AccessControlCondition
@@ -87,26 +88,8 @@ import {
  */
 export default class LitNodeClient {
   constructor(config) {
-    this.config = {
-      alertWhenUnauthorized: true,
-      minNodeCount: 6,
-      debug: true,
-      bootstrapUrls: [
-        "https://node2.litgateway.com:7370",
-        "https://node2.litgateway.com:7371",
-        "https://node2.litgateway.com:7372",
-        "https://node2.litgateway.com:7373",
-        "https://node2.litgateway.com:7374",
-        "https://node2.litgateway.com:7375",
-        "https://node2.litgateway.com:7376",
-        "https://node2.litgateway.com:7377",
-        "https://node2.litgateway.com:7378",
-        "https://node2.litgateway.com:7379",
-      ],
-    };
-    if (config) {
-      this.config = { ...this.config, ...config };
-    }
+    // configure() also stores config to `globalThis.litConfig`, see function def
+    this.config = configure(config);
 
     this.connectedNodes = new Set();
     this.serverKeys = {};
@@ -114,20 +97,6 @@ export default class LitNodeClient {
     this.subnetPubKey = null;
     this.networkPubKey = null;
     this.networkPubKeySet = null;
-
-    try {
-      if (typeof window !== "undefined" && window && window.localStorage) {
-        let configOverride = window.localStorage.getItem("LitNodeClientConfig");
-        if (configOverride) {
-          configOverride = JSON.parse(configOverride);
-          this.config = { ...this.config, ...configOverride };
-        }
-      }
-    } catch (e) {
-      console.log("Error accessing local storage", e);
-    }
-
-    globalThis.litConfig = this.config;
   }
 
   /**
